moving OAuthTokenInfo Class to api package and renaming APIManagerJWTAuthComponent class
